+++ packages/gamut-labs/src/AppHeader/AppHeaderElements/AppHeaderCatalogDropdown/index.tsx
@@ -238,6 +238,7 @@ export const AppHeaderCatalogDropdown: React.FC<AppHeaderCatalogDropdownProps> =
           keyDownEvents={menuHandleKeyEvents}
           id={`menu-container${item.text}`}
           isOpen={isOpen}
+          handleClose={() => setIsOpen(false)}
         />
       </StyledDropdown>
     </>