+++ packages/gamut/src/Modals/__tests__/Modal-test.tsx
@@ -8,9 +8,7 @@ import { Modal, ModalProps } from '..';
 const renderModal = (props?: Partial<ModalProps>) => {
   return render(
     <ThemeProvider theme={theme}>
-      <Modal isOpen onRequestClose={jest.fn()} {...props}>
-        <div data-testid="modal-content">{props?.children}</div>
-      </Modal>
+      <Modal isOpen onRequestClose={jest.fn()} {...props} />
     </ThemeProvider>
   );
 };