+++ packages/gamut/src/index.tsx
@@ -49,6 +49,7 @@ export * from './ProgressBar';
 export * from './RadialProgress';
 export * from './SkipToContent';
 export * from './Spinner';
+export * from './Tag';
 export * from './Tabs';
 export * from './theming/VisualTheme';
 export * from './Toast/Toast';