+++ packages/styleguide/stories/Foundations/ColorMode/examples.tsx
@@ -229,6 +229,31 @@ export const BackgroundExample: React.FC<ComponentProps<typeof Background>> = ({
   );
 };
 
+export const SystemColorModeExample = () => {
+  const prefersDarkMode = usePrefersDarkMode();
+  const mode = prefersDarkMode ? 'dark' : 'light';
+  const bg = prefersDarkMode ? 'navy' : 'paleBlue';
+
+  return (
+    <ColorMode mode="system">
+      <Background bg={bg} p={24}>
+        <Text variant="title-md" mb={16}>
+          system color mode:{' '}
+          <Text color="primary" fontFamily="monospace" ml={8}>
+            {mode} mode
+          </Text>
+        </Text>
+        <Text as="p" mb={16}>
+          Lorem ipsum dolor sit amet, sed do eiusmod tempor incididunt ut labore
+          et dolore <Anchor>magna aliqua</Anchor>. Ut <a href="#cool">enim</a>{' '}
+          ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+          aliquip ex ea commodo consequat.
+        </Text>
+      </Background>
+    </ColorMode>
+  );
+};
+
 export const PrefersDarkModeExample = () => {
   const prefersDarkMode = usePrefersDarkMode();
   const mode = prefersDarkMode ? 'dark' : 'light';